// FrontApp/src/app/model/client.model.ts
export class Client {
  constructor() {
  }

  id?: number;
  organisation?: string;
  codePostal?: string;
  adressSiege?: string;
  adressLiv?: string;
  telMobile?: string;
  ville?: string;
  pays?: string;
  telFixe?: string;
  telFaxe?: string;
  email?: string;
  raisonSoc?: string;
  statut?: string;
  urlSoc?: string;
  fonction?: string;
  type?: string;
  numContrat?: string;
  chefProjet?: string;
//liste de contacts

}